IR phantom check (#260)
* boilerplate

* gather events

* more things

* invoke

* need to fix

* add interface info for all

* non-fails passing now

* add phantom tests

* remove option

* remove comments

* forgot some

* more comments oops

* change back to old interfaces

* remove unused fn

* remove unused field

* misc

* change to old funcs

* remove default

* comment

* undo unneeded changes

* restructure to allow multiple errors and loop check

* update tests

* test reporting multiple phantom errors

---------

+++ src/ir_passes/assignment_check.rs
@@ -18,7 +18,7 @@ use linked_hash_map::LinkedHashMap;
 /// Must occur after monomorphization.
 pub struct AssignCheck {
     ports: LinkedHashMap<(PortIdx, usize), Vec<Option<GPosIdx>>>,
-    diagnostic_count: u32,
+    diagnostic_count: u64,
 }
 
 impl Construct for AssignCheck {
@@ -152,7 +152,7 @@ impl Visitor for AssignCheck {
         }
     }
 
-    fn after_traversal(&mut self) -> Option<u32> {
+    fn after_traversal(&mut self) -> Option<u64> {
         if self.diagnostic_count > 0 {
             Some(self.diagnostic_count)
         } else {
